+++ lib/elixir/lib/option_parser.ex
@@ -111,7 +111,7 @@ defmodule OptionParser do
     Keyword.put dict, option, value
   end
 
-  defp normalize_option(<<?-, option|:binary>>, aliases) do
+  defp normalize_option(<<?-, option :: binary>>, aliases) do
     normalize_option(option, aliases)
   end