+++ lib/elixir/test/elixir/binary_test.exs
@@ -88,19 +88,19 @@ bar
 
   test :pattern_match do
     s = 16
-    assert <<a, b|s>> = "foo"
+    assert <<a, b :: size(s)>> = "foo"
   end
 
   test :bitsyntax_translation do
     refb = "sample"
     sec_data = "another"
-    << size(refb) | 1 - :big - :signed - :integer - {:unit, 8},
-       refb | :binary,
-       size(sec_data) | 1 - :big - :signed - :integer - {:unit, 16},
-       sec_data|:binary >>
+    << size(refb) :: [size(1), big, signed, integer, unit(8)],
+       refb :: binary,
+       size(sec_data) :: [size(1), big, signed, integer, unit(16)],
+       sec_data :: binary >>
   end
 
-  defp is_match?(<<char, _|:binary>>, char) do
+  defp is_match?(<<char, _ :: binary>>, char) do
     true
   end