+++ lib/elixir/test/elixir/string/inspect_test.exs
@@ -50,7 +50,7 @@ defmodule String.Inspect.BitStringTest do
   use ExUnit.Case, async: true
 
   test :bitstring do
-    assert inspect(<<1|12-:integer-:signed>>) == "<<0,1|4>>"
+    assert inspect(<<1 :: [size(12), integer, signed]>>) == "<<0,1|4>>"
   end
 
   test :binary do